[PRERELEASE] LLM integration, dynamic model selection, markdown rendering, and react routing (#5)
* merge: receive streaming response from language model server API (#2)

This pull request add basic React hooks to receive a streaming message from the language server and display it in the chat window. The text is unformatted but animates in as each chunk of content is received from the server.

* feat: add hooks for fetching data and getting streaming chat data
* feat: remove dummy text, create states, pass changes from input to chat box for handling
* feat: handle interaction (click, submit)
* feat: receive single, animated response from prompt
* refactor: remove unused components and imports
* style: fix code style
* docs: update changelog

* feat: markdown parsing/rendering in chat (#3)

Markdown provides a structured way to handle the output from the LLM. The application now renders the markdown from the LLM and converts it into styled HTML and React components. This provides a dynamic chat experience where sections are divided into separate chat bubbles and headings, titles, and text decorations are all dynamically styled.

* feat: parse headings, paragraphs, and lists into single bubble
* feat: parse some inline characters
* feat: render markdown acceptably well
* fix: bullets that start with formatted text now on same line
* chore: fix node being out of date
* docs: update readme and changelog

* feat: add sidebar routing, model selection (#4)

This pull request includes several changes to the frontend and development environment. The main focus is on integrating React Router for navigation, adding a model selection feature, and updating dependencies.

* feat: create components for sidebar menu; sync animations on expanding and hover
* chore: fix invalid dockerfile; add react router
* feat: add menu items and react router navigation
* refactor: adjust margins to universally align main content outside of sidebar space
* feat: add blank homepage
* feat: basic homepage
* feat: update homepage style color and spacing
* feat: select a model and send a message with that model

* docs: update readme and changelog